wgipam: split Lease into new file
