+++ store.go
@@ -15,10 +15,7 @@ package wgipam
 
 import (
 	"encoding/binary"
-	"encoding/json"
-	"fmt"
 	"io"
-	"net"
 	"sync"
 	"time"
 
@@ -30,42 +27,6 @@ var (
 	_ Store = &memoryStore{}
 	_ Store = &boltStore{}
 )
-
-// A Lease is a record of allocated IP addresses assigned to a client.
-type Lease struct {
-	IPv4, IPv6 *net.IPNet
-	Start      time.Time
-	Length     time.Duration
-}
-
-// String returns a string suitable for logging.
-func (l *Lease) String() string {
-	return fmt.Sprintf("IPv4: %s, IPv6: %s, start: %s, end: %s",
-		l.IPv4, l.IPv6,
-		// time.Stamp seems to be reasonably readable.
-		l.Start.Format(time.Stamp),
-		l.Start.Add(l.Length).Format(time.Stamp),
-	)
-}
-
-// Expired determines if the Lease is expired as of time t.
-func (l *Lease) Expired(t time.Time) bool {
-	exp := l.Start.Add(l.Length)
-	return exp.Before(t) || exp.Equal(t)
-}
-
-// TODO(mdlayher): JSON is quick and easy but it's probably best to build out
-// a binary format.
-
-// marshal marshals a Lease to binary form.
-func (l *Lease) marshal() ([]byte, error) {
-	return json.Marshal(l)
-}
-
-// unmarshal unmarshals a Lease from binary form.
-func (l *Lease) unmarshal(b []byte) error {
-	return json.Unmarshal(b, l)
-}
 
 // A Store manages Leases. To ensure compliance with the expected behaviors
 // of the Store interface, use the wgipamtest.TestStore function.
